Add files via upload
This file contains an attempt to create a personalized GPT for a smaller set of data that includes 6 varied judgements of different types. Based on this, any number of questions could be answered. Along with this, there has also been mapping of text, that is, on giving the detail of a case, its corresponding relevant section number of the Indian Penal Code is shown as output.